précision localisation

+++ app/src/main/java/fr/android/projet_polart_masbernat/SecondActivity.java
@@ -375,8 +375,8 @@ public class SecondActivity extends AppCompatActivity implements LocationListene
 
     @Override
     public void onLocationChanged(Location location) {
-        int lat = (int) (location.getLatitude());
-        int lng = (int) (location.getLongitude());
+        float lat = (float) (location.getLatitude());
+        float lng = (float) (location.getLongitude());
         latituteField.setText(String.valueOf(lat));
         longitudeField.setText(String.valueOf(lng));
     }
